author tf idf

+++ corpus/utils.py
@@ -1,5 +1,9 @@
 import os
 import shutil
+import json
+import tqdm
+import re
+from prompt_toolkit import prompt
 
 from nltk.stem.snowball import SnowballStemmer
 from gensim import corpora
@@ -22,23 +26,24 @@ def build_keys(keys: list):
     return [tuple(k.split()) for k in keys]
 
 
-def num_dict(year_list: list, keywords: [list, None]=None, nested: [int, None]=0):
+def num_dict(alist: list, keywords: [list, None]=None, nested: [int, None]=0):
     """
     Build empty dictionary with integers at leaf entries.
+    alist: year_list or author_list
     """
 
     results = {}
 
-    for year in year_list:
+    for item in alist:
 
         if nested == 0:
-            results[year] = 0
+            results[item] = 0
 
         elif nested == 1:
-            results[year] = {}
-            results[year]['TOTAL'] = 0
+            results[item] = {}
+            results[item]['TOTAL'] = 0
             for keyword in keywords:
-                results[year][keyword] = 0
+                results[item][keyword] = 0
 
         else:
             _fail('Shouldn\'t be able to get here.')
@@ -46,23 +51,24 @@ def num_dict(year_list: list, keywords: [list, None]=None, nested: [int, None]=0
     return results
 
 
-def list_dict(year_list: list, keywords: [list, None]=None, nested: [None, int]=0):
+def list_dict(alist: list, keywords: [list, None]=None, nested: [None, int]=0):
     """
     Build empty dictionary with lists at leaf entries.
+    alist: year_list or author_list
     """
 
     results = {}
 
-    for year in year_list:
+    for item in alist:
 
         if nested == 0:
-            results[year] = []
+            results[item] = []
 
         elif nested == 1:
-            results[year] = {}
+            results[item] = {}
             for keyword in keywords:
-                results[year]['TOTAL'] = []
-                results[year][keyword] = []
+                results[item]['TOTAL'] = []
+                results[item][keyword] = []
 
         else:
             _fail('Shouldn\'t be able to get here.')
@@ -70,15 +76,16 @@ def list_dict(year_list: list, keywords: [list, None]=None, nested: [None, int]=
     return results
 
 
-def gensim_dict(year_list: list):
+def gensim_dict(alist: list):
     """
     Build empty dictionary with gensim Dictionary objects at leaf entries.
+    alist: year_list or author_list
     """
 
     results = {}
 
-    for year in year_list:
-        results[year] = corpora.Dictionary()
+    for item in alist:
+        results[item] = corpora.Dictionary()
 
     return results
 
@@ -112,7 +119,7 @@ def stem(word: str, language: [str, None] = 'english'):
     return '{0}: {1}'.format(word, stemmed)
 
 
-# TODO: ask user if they want to overwrite directory or add to it, if it exists
+# TODO: ask user if they want to overwrite directory or add to it, if it exists --DONE
 def build_out(out_dir: str):
     """
     Build output directory, overwrite if it exists.
@@ -122,13 +129,23 @@ def build_out(out_dir: str):
         if not os.path.exists(out_dir):
             os.mkdir(out_dir)
         else:
-            shutil.rmtree(out_dir)
-            os.mkdir(out_dir)
+            while True:
+                ans = prompt("\nDirectory already exists. Do you want to overwrite[O] or add[A] to it?    ")
+                if ans == 'O':
+                    print("overwrite...")
+                    shutil.rmtree(out_dir)
+                    os.mkdir(out_dir)
+                    break
+                elif ans == 'A':
+                    print("add to...")
+                    # do nothing
+                    break
+                else:
+                    print("Command not recognized.")
+                    continue
+
     else:
         _fail("Please specify output directory.")
-
-
-
